Updated device_id to use the actual VIN
Suggesting a change to the identifier for a vehicle away from a UUID and using the operators identifier instead.   The thought here is that each bike manufacturer already has a unique identifier for their vehicle embedded in their code (QR code or otherwise).  It's part of the inherent of the business.   Let's use that number so they could be correlated to real bikes on the street if it ever became necessary.