Require Standard HTTP Status Codes for Provider
I initially proposed #113, which was correctly pointed out as a dupe of #105. However, looking at the relevant PR #156 , it only addresses the agency spec without also addressing Provider. This adds clarification around standard HTTP status codes for the provider part of #105.  I added it under Response Format, but open to other ideas if there is a better placement elsewhere.