Support batch telemetry
The submission of telemetry data is now provided on `POST /vehicles/telemetry` to allow for batch telemetry data upload.  It will still support single telemetry point submission, but also across an entire fleet of active vehicles or data for an entire trip.  The goal is to make it easier and more flexible for Providers to submit data and to design to support larger scale datasets.  Telemetry data no longer requires a trip_id and will be automatically associated with trips based on timestamps.

The telemetry data for this endpoint is now structured as a simple JSON object and not GeoJSON.  GeoJSON will remain for the GET (retrieval) endpoints.