Add dynamic properties for service areas
Adding the concept of a properties object for service_areas.  The idea is that this will provide an extensible way to attach dynamic properties to returned service areas that is specific to the Provider.  The first proposal is a vehicle quota to let Providers know how many vehicles they can deploy in service areas available to them.