Require each event to inclue a UUID.
Per discussion in #191, while in theory every event should be
unique for every `(provider_id, device_id, event_time)` tuple,
in practice this is not the case, for a variety of reasons. To
help facilitate data ingestion, referential integrity, and general
debugging/discussion, we're adding a required UUID column to every
status_change event.

Fixes #191