Allow any status change to include an `associated_trip`
PR #297 already updated the spec language to make it clear that any
status change is allowed to have an `associated_trip`, but I didn't
realize this was incompatible with the existing schema. The schema
should now match the spec, which states that `user_pick_up` and
`user_drop_off` events must have an `associated_trip` but other events
may have one when applicable.