provider: Make Trip's 'route' field optional.
For many regulatory and privacy-oriented reasons, it's possible that
cities or providers do not want external actors to access the
geolocalized data of the trips throughout the city.

The distance, average cost, speed, duration of those trips is still
valuable and aggregated information and should thus be accessible, even
if the route may not be.