Clarify policy rule min and max are inclusive
The policy docs don't say anywhere whether the minimum and
maximum values specified on policy rules are to be interpretted
as a closed or open interval (that is, whether the endpoints are
included in the interval). Most real world policies seem to treat
these min and maxes inclusively, e.g. parking for a max of
2 hours or a 70 MPG speed limit. You're not in violation until you
go over those maximums. So this change clarifies that min and max
values on policy rules are meant to define and inclusive (or closed)
interval where values equal to the min and max match the rule.