Add a rate_applies_when field to Policy Rules
The rate_applies_when field allows people constructing policy
rules with attached rates to specify when the rate is applicable:
is when something is within the rule bounds or when it is outside
of them? The rate_applies_when when field may take on the values
in_bounds and out_of_bounds.