Add trip definition to general information file
This addresses #748 by adding a definition of a "trip" to the
general-information.md file. The definition is meant for the
purposes of analysis in particular, and not for filtering which
data points are exchanged as part of MDS endpoints.

I wasn't sure on the exact right spot for this since I hadn't seen
anything like this done before in MDS, but I'm open to suggestions
on that and the wording of the definition itself.