Add CDS Rate object to MDS
When developing CDS we came up with a more expansive definition of
rates [1] to better capture real life scenarios. This commit brings
that same definition to MDS, but keeps the MDS field rate_applies_when
for specifying when the rate applies to an event.

As in CDS, a policy rule may now contain an array of rates specifying
monetary charges or subsidies for events that match that rule.
This removes the need to create multiple rules on a policy to capture
different parts of the rate calculation, it can all be contained in the
rate array.

[1]: https://github.com/openmobilityfoundation/curb-data-specification/tree/1.0.0/curbs#rate